plugin/tongkongyuan/python/requirements.txt
absl-py==0.9.0
addict==2.2.1
albumentations==0.4.3
alembic==1.4.0
apex==0.1
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
boto3==1.12.39
botocore==1.15.39
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
configparser==4.0.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
databricks-cli==0.9.1
dataclasses==0.7
davisinteractive==0.0.1.dev9
decorator==4.4.0
defusedxml==0.6.0
docker==4.2.0
docutils==0.15.2
dominate==2.5.1
easydict==1.9
efficientnet-pytorch==0.6.3
en-core-web-sm==2.2.5
entrypoints==0.3
fastparquet==0.3.2
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gast==0.3.3
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
grpcio==1.26.0
h5py==2.10.0
hyperopt==0.1.2
idna==2.8
imagededup==0.2.2
imageio==2.5.0
imgaug==0.2.6
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iterative-stratification==0.1.6
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.13.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
librosa==0.7.2
lightgbm==2.3.1
llvmlite==0.30.0
Mako==1.1.1
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mlflow==1.6.0
mmcv==0.2.15
more-itertools==7.2.0
munch==2.3.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.3
nlpaug==0.0.14
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.46.0
numpy==1.16.6
numpy-quaternion==2019.10.3.10.26.21
oauthlib==3.1.0
olefile==0.46
onnx==1.7.0
opencv-python==4.2.0.32
opencv-python-headless==4.1.1.26
opt-einsum==3.2.1
pandas==1.0.5
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
plac==1.1.3
portalocker==1.7.0
pot==1.0
prefetch-generator==1.0.1
preshed==3.0.2
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prometheus-flask-exporter==0.12.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.7.0
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
pypiwin32==223
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.5
python-dateutil==2.8.0
python-editor==1.0.4
pytorch-lightning==0.8.4
pytorchcv==0.0.57
pytz==2019.2
PyWavelets==1.0.3
pywin32==225
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
querystring-parser==1.2.4
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
s3transfer==0.3.3
sacrebleu==1.4.9
sacremoses==0.0.38
scikit-image==0.15.0
scikit-learn==0.23.1
scikit-multilearn==0.2.0
scipy==1.3.1
seaborn==0.9.0
segmentation-models-pytorch==0.1.0
Send2Trash==1.5.0
sentencepiece==0.1.85
SimpleITK==1.2.4
simplejson==3.17.0
six==1.12.0
smmap2==2.0.5
SoundFile==0.10.3.post1
spacy==2.2.4
SQLAlchemy==1.3.13
sqlparse==0.3.0
srsly==1.0.2
tabulate==0.8.6
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
theconf==0.1.3
thinc==7.4.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
thrift==0.13.0
timm==0.2.1
tokenizers==0.7.0
torch==1.5.1+cu101
torch-encoding==1.0.1
torchfile==0.1.0
torchnet==0.0.4
torchstat==0.0.7
torchvision==0.6.1+cu101
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
transformers==2.10.0
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.3
visdom==0.1.8.8
waitress==1.4.3
wasabi==0.6.0
wcwidth==0.1.7
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.1.1
xlrd==1.2.0
zipp==0.6.0
